+++ front-end/src/renderer/pages/RestoreMissingKeys/RestoreMissingKeys.vue
@@ -16,7 +16,7 @@ import {
 } from '@renderer/utils';
 
 import AppButton from '@renderer/components/ui/AppButton.vue';
-import Import from '@renderer/pages/AccountSetup/components/Import.vue';
+import Import from '@renderer/components/RecoveryPhrase/Import.vue';
 
 /* Stores */
 const user = useUserStore();
